Add support to create 'iperf3' binary in Android.bp
Tools like sigma-dut expects iperf3 tool name to be "iperf3" but
Android.bp was generating it as "iperf"

This commit adds support to generate binary as iperf3 by adding
new "cc_binary" block with module name as iperf3.

Also the support for generating the module as "iperf" is unmodified
to be compatible with default behaviour.

Change-Id: Id85fd14afe05605e64e3667555caf546232a4404
CRs-Fixed: 2616760